Moved writing to final stage instead of during the compression

+++ issue1/src/utils/include/utilities.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "../../cliparser/include/cli.h"
 #include <stdio.h>
-int readIntSamples(struct arguments * parameters, char fileName[128], uint32_t * samples);
+int readIntSamples(struct arguments * parameters, char fileName[128], uint16_t * samples);
 void printArrayLong(unsigned long * sample, struct arguments * args);
 double walltime ( void );
 void writeArrayLong(unsigned long * sample, FILE * file, struct arguments * args);
